requirements.txt
adblockparser
arrow
celery
Django<1.12
djangorestframework
django-widget-tweaks
dnspython
geoip2
msgpack-python
Pillow
psycopg2 # required on slaves without db access as well due to django.contrib.postgres imports in models
redis
requests
tldextract
toposort
url_normalize
pygments
django-pandas
matplotlib
mpld3
orange3
orange3-associate
gunicorn
django-debug-toolbar
fancyimpute
seaborn
